# docs/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Hotel competitor monitor API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/pms/otb/template": {
   "get": {
    "tags": [
     "otb"
    ],
    "summary": "Otb Template",
    "operationId": "otb_template_pms_otb_template_get",
    "parameters": [
     {
      "name": "kind",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "otb_report",
        "bookings"
       ],
       "type": "string",
       "default": "otb_report",
       "title": "Kind"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "text/plain": {
        "schema": {
         "type": "string"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/pms/otb/import": {
   "post": {
    "tags": [
     "otb"
    ],
    "summary": "Import Otb",
    "description": "Nhập OTB không ghi đè ngày khác: (khách sạn, ngày chụp, đêm) là khoá. File đặt phòng dựng\nlại bản chụp cho cả quá khứ (pace/STLY có ngay).",
    "operationId": "import_otb_pms_otb_import_post",
    "parameters": [
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_import_otb_pms_otb_import_post"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OtbImportOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/market/otb": {
   "get": {
    "tags": [
     "otb"
    ],
    "summary": "Otb View",
    "description": "OTB, pickup, pace (4 tuần trước, STLY) và dự báo theo đêm của khách sạn của bạn.",
    "operationId": "otb_view_market_otb_get",
    "parameters": [
     {
      "name": "start",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "Start"
      }
     },
     {
      "name": "end",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "format": "date"
        },
        {
         "type": "null"
        }
       ],
       "title": "End"
      }
     },
     {
      "name": "own_hotel_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Own Hotel Id"
      }
     },
     {
      "name": "days",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 120,
       "minimum": 1,
       "default": 30,
       "title": "Days"
      }
     },
     {
      "name": "tenant_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Operator: tenant cần xem",
       "title": "Tenant Id"
      },
      "description": "Operator: tenant cần xem"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OtbOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_import_otb_pms_otb_import_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     },
     "hotel_id": {
      "type": "integer",
      "title": "Hotel Id"
     },
     "kind": {
      "type": "string",
      "enum": [
       "otb_report",
       "bookings"
      ],
      "title": "Kind",
      "default": "otb_report"
     },
     "as_of_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "As Of Date"
     },
     "rooms_available": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rooms Available"
     }
    },
    "type": "object",
    "required": [
     "file",
     "hotel_id"
    ],
    "title": "Body_import_otb_pms_otb_import_post"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "KpiOut": {
    "properties": {
     "nights": {
      "type": "integer",
      "title": "Nights"
     },
     "rooms_sold": {
      "type": "integer",
      "title": "Rooms Sold"
     },
     "rooms_available": {
      "type": "integer",
      "title": "Rooms Available"
     },
     "revenue": {
      "type": "string",
      "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
      "title": "Revenue"
     },
     "occupancy_pct": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Occupancy Pct"
     },
     "adr": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Adr"
     },
     "revpar": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Revpar"
     }
    },
    "type": "object",
    "required": [
     "nights",
     "rooms_sold",
     "rooms_available",
     "revenue",
     "occupancy_pct",
     "adr",
     "revpar"
    ],
    "title": "KpiOut"
   },
...
   "OtbImportOut": {
    "properties": {
     "kind": {
      "type": "string",
      "title": "Kind"
     },
     "hotel_id": {
      "type": "integer",
      "title": "Hotel Id"
     },
     "rows_read": {
      "type": "integer",
      "title": "Rows Read"
     },
     "snapshots_written": {
      "type": "integer",
      "title": "Snapshots Written"
     },
     "as_of_dates": {
      "type": "integer",
      "title": "As Of Dates"
     },
     "stay_dates": {
      "type": "integer",
      "title": "Stay Dates"
     },
     "errors": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "type": "array",
      "title": "Errors"
     },
     "status": {
      "type": "string",
      "title": "Status"
     }
    },
    "type": "object",
    "required": [
     "kind",
     "hotel_id",
     "rows_read",
     "snapshots_written",
     "as_of_dates",
     "stay_dates",
     "errors",
     "status"
    ],
    "title": "OtbImportOut"
   },
   "OtbNightOut": {
    "properties": {
     "stay_date": {
      "type": "string",
      "format": "date",
      "title": "Stay Date"
     },
     "lead": {
      "type": "integer",
      "title": "Lead"
     },
     "rooms_otb": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rooms Otb"
     },
     "revenue_otb": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Revenue Otb"
     },
     "rooms_available": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rooms Available"
     },
     "occ_otb_pct": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Occ Otb Pct"
     },
     "adr_otb": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Adr Otb"
     },
     "pickup_1d": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pickup 1D"
     },
     "pickup_7d": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pickup 7D"
     },
     "pace_4w": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pace 4W"
     },
     "ref_4w": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ref 4W"
     },
     "stly": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Stly"
     },
     "pace_stly": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pace Stly"
     },
     "forecast_rooms": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Forecast Rooms"
     },
     "forecast_occ_pct": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Forecast Occ Pct"
     },
     "forecast_basis": {
      "type": "integer",
      "title": "Forecast Basis"
     }
    },
    "type": "object",
    "required": [
     "stay_date",
     "lead",
     "rooms_otb",
     "revenue_otb",
     "rooms_available",
     "occ_otb_pct",
     "adr_otb",
     "pickup_1d",
     "pickup_7d",
     "pace_4w",
     "ref_4w",
     "stly",
     "pace_stly",
     "forecast_rooms",
     "forecast_occ_pct",
     "forecast_basis"
    ],
    "title": "OtbNightOut"
   },
   "OtbOut": {
    "properties": {
     "hotel_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Hotel Id"
     },
     "as_of_date": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "As Of Date"
     },
     "start": {
      "type": "string",
      "format": "date",
      "title": "Start"
     },
     "end": {
      "type": "string",
      "format": "date",
      "title": "End"
     },
     "nights": {
      "items": {
       "$ref": "#/components/schemas/OtbNightOut"
      },
      "type": "array",
      "title": "Nights"
     },
     "actual_30d": {
      "$ref": "#/components/schemas/KpiOut"
     },
     "otb_window": {
      "$ref": "#/components/schemas/KpiOut"
     },
     "price_index_median": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Price Index Median"
     },
     "price_index_nights": {
      "type": "integer",
      "title": "Price Index Nights"
     }
    },
    "type": "object",
    "required": [
     "hotel_id",
     "as_of_date",
     "start",
     "end",
     "nights",
     "actual_30d",
     "otb_window",
     "price_index_median",
     "price_index_nights"
    ],
    "title": "OtbOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}